lib: extract client-safe enums (UserRole, TaskStatus, ENTITY_TYPES)
Pulls UserRole, TaskStatus, and ENTITY_TYPES into ~/lib/ so client
components can import them without dragging Drizzle into the browser
bundle via the ~/server/db/schema barrel (TSS-2 rule).

The schema files (auth.schema.ts, task.schema.ts) and ability.ts now
import from ~/lib/ and re-export, so existing server code that imports
from ~/server/db/schema keeps working unchanged. Client-side imports
(components, routes, validation.ts, playwright.ts) are flipped to
~/lib/ directly.

Verified by rebuilding and grepping the client bundle:
  grep -lE 'postgres-js|"drizzle:entityKind"' .output/public/assets/main-*.js
returns nothing. The eager client graph is now Drizzle-free; remaining
hits in lazy Better Auth adapter chunks (kysely-adapter, dialect,
migrator) are vendor code-split that isn't loaded by the main bundle.
