+++ src/components/admin/UserForm.tsx
@@ -1,5 +1,5 @@
-import type { User, UserRoleType } from "~/server/db/schema";
-import { UserRole } from "~/server/db/schema";
+import { UserRole, type UserRoleType } from "~/lib/enums/user-role";
+import type { User } from "~/server/db/schema";
 import { userFormSchema } from "~/types/validation";
 import { FormGenerator } from "../form/FormGenerator";
 import type { FormFieldConfig } from "../form/types";
